refact(address): add lot and property routes to address controller
